FP6: overlay: Refine display cutout region
By observing the screen with a macro camera, we can observe that the
cutout has a x radius of 38px, a y radius of 39px and begins 25px below
the top border. The rectangle approximation of the cutout should have
the same amount of padding on top of the cutout as below according to
the pixel device trees ("This configuartion effects status bar's height
actually. Set its height as double y position of
@dimen/config_mainBuiltInDisplayCutout to make cutout center
vertically" [1]). Additionally, add 8px padding at the left and right
side of the cutout to avoid content being too close to the cutout.
Below is a sketch of the cutout drawn within the rectangle
approximation.
___________
|         |
|    _    |  ___25px
|  /   \  |
| |     | |
|  \   /  |
|    ‾    |  ‾‾‾103x
|         |
‾‾‾‾‾‾‾‾‾‾‾  ‾‾‾128px
  |     | |
  8px   | |
        | |
        84px
          |
          92px

[1]: https://android.googlesource.com/device/google/caimito/+/refs/heads/main/caiman/overlay/frameworks/base/core/res/res/values/config.xml#197

Change-Id: I117a347bbaddc42185b31127733fa7ee5078d147