FP6: Set debug.camera.packagename to com.fps.camera
The 50MP High Pixel mode is fixed for use by other camera apps without
causing crashes in CheckValidStreamConfig(). The property is set to
utilize the bypass present in the CamX code, which was identified
through reverse-engineering. It is returned as having a valid stream
config when the caller is com.fps.camera, Fairphone's Camera app.

Therefore hardcode the property to the same to utilize the bypass on
other camera applications as well.

Co-authored-by: Rohit R Sekhar <merothh@e.email>
Change-Id: I94b965645be0d89b200040136a3b48ecbdba2fcc